+++ packages/frontend/src/App.css
@@ -12,6 +12,7 @@
   background-color: white;
   min-height: 100%;
   min-width: 50%;
+  max-width: 50%;
   display: flex;
   flex-direction: column;
   align-items: center;
